\name{wowa}
\alias{wowa}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{WOWA package
}
\description{Various weighted multivariate extensions of bivariate and OWA functions, including implicit, quantifier-based and binary tree based WOWA.
}
\usage{
  wowa()
}
\details{Lists the functions implemented in this package.
}  
  \value{
    \item{output}{No return value, called for printing only.}
  }


  \references{
    %% ~put references to the literature/web site here ~
[1]G. Beliakov, H. Bustince, and T. Calvo. A Practical Guide to Averaging Functions.
	Springer, Berlin, Heidelberg, 2016.

[2]G. Beliakov. A method of introducing weights into OWA operators and other
	symmetric functions. In V. Kreinovich, editor, Uncertainty Modeling. Dedicated
	to B. Kovalerchuk, pages 37-52. Springer, Cham, 2017.

[3]G. Beliakov. Comparing apples and oranges: The weighted OWA function, Int.J. Intelligent Systems, 33, 1089-1108, 2018.

[4]V. Torra. The weighted OWA operator. Int. J. Intelligent Systems, 12:153-166, 1997.

[5]G. Beliakov and J.J. Dujmovic , Extension of bivariate means to weighted means of several arguments by using binary trees, 
Information sciences, 331,  137-147,  2016.

[6]	J.J. Dujmovic and G. Beliakov. Idempotent weighted aggregation based on binary
	aggregation trees. Int. J. Intelligent Systems  32,  31-50,  2017.
  }
  \author{
    %%  ~~who you are~~
      Gleb Beliakov, Daniela L. Calderon, Deakin University
  }
  
  \examples{
	wowa()
  }
  % Add one or more standard keywords, see file 'KEYWORDS' in the
  % R documentation directory.
  \keyword{ wowa}
  \keyword{WOWA}% __ONLY ONE__ keyword per line